backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0
google-api-python-client==2.118.0
google-genai>=1.0.0
bleach==6.1.0